requirements.txt
anyio==4.10.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
feedparser==6.0.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
notion-client==2.4.0
python-dotenv==1.1.1
requests==2.32.4
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
